chore: Split bot into multiple files
