+++ .github/api-extractor-action/src/const.ts
@@ -1,0 +1,3 @@
+export const BASE_BRANCH_DIR = 'branch-clone';
+export const BUILD_DIR = 'dist';
+export const REPORT_DIR = 'etc';
